+++ src/Entity/StationMedia.php
@@ -6,7 +6,6 @@ namespace App\Entity;
 
 use App\Annotations\AuditLog;
 use App\Flysystem\FilesystemManager;
-use App\Media\Metadata;
 use App\Normalizer\Annotation\DeepNormalize;
 use Doctrine\Common\Collections\ArrayCollection;
 use Doctrine\Common\Collections\Collection;